refactor: Mejora el naming de los archivos siguiendo la convención de Next.js
