+++ components/shared/highlighted-text/highlighted-text.tsx
@@ -1,4 +1,4 @@
-import StarWithStroke from './starWithStroke';
+import StarWithStroke from '../star-with-stroke/star-with-stroke';
 
 interface HighlightedTextProps {
   svgSize?: number;
