+++ components/shared/star-with-stroke/star-with-stroke.tsx
