+++ components/shared/stars/index.ts
