+++ lib/utils/is-defined.ts
